feat: add float_format flag
This commit adds a float_format flag to easyjson which allows
users to change the float format used in strconv. The default flag
used is `g` by easyjson.

If a user now provides a flag such as:
	bin/easyjson -float_format='f' ./tests/float_format.go

It will use that format. The g format converts long floats into the
e-notation. Using a flag such as 'f' might be necessary in certain
usecases to maintain backward compatibility.